state_manager: Fix unit tests publishing late events
This CL distinguishes between MorseEncodeRequests and TimerRequests
Published by the StateManager. Previously a single thread notification
was used for both, and as a result tests did not wait for the
TimerRequests to be sent. This led them to be sent after the worker was
stopped.

Bug: 357935830
Change-Id: Ifef907c8d3f66cee391e122ceeef068a4f256452
Reviewed-on: https://pigweed-internal-review.git.corp.google.com/c/pigweed/showcase/rp2/+/74093
Commit-Queue: Aaron Green <aarongreen@google.com>
Reviewed-by: Wyatt Hepler <hepler@google.com>
Presubmit-Verified: CQ Bot Account <pigweed-internal-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Ted Pudlik <tpudlik@google.com>